@@ -10,12 +10,12 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_02_23_014118) do
+ActiveRecord::Schema[7.0].define(version: 2023_09_13_215307) do
   create_table "movies", force: :cascade do |t|
     t.string "title"
     t.string "rating"
     t.text "description"
-    t.string "release_date"
+    t.datetime "release_date"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
