@@ -5,20 +5,21 @@
 #
 #   movies = Movie.create([{ name: "Star Wars" }, { name: "Lord of the Rings" }])
 #   Character.create(name: "Luke", movie: movies.first)
+
 # Seed the RottenPotatoes DB with some movies.
 more_movies = [
-  {:title => 'My Neighbor Totoro', :rating => 'G',
-    :release_date => '16-Apr-1988'},
-  {:title => 'Green Book', :rating => 'PG-13',
-    :release_date => '16-Nov-2018'},
-  {:title => 'Parasite', :rating => 'R',
-    :release_date => '30-May-2019'},
-  {:title => 'Nomadland', :rating => 'R',
-    :release_date => '19-Feb-2021'},
-  {:title => 'CODA', :rating => 'PG-13',
-    :release_date => '13-Aug-2021'}
-]
-
-more_movies.each do |movie|
-  Movie.create!(movie)
-end
+    {:title => 'My Neighbor Totoro', :rating => 'G',
+      :release_date => '16-Apr-1988'},
+    {:title => 'Green Book', :rating => 'PG-13',
+      :release_date => '16-Nov-2018'},
+    {:title => 'Parasite', :rating => 'R',
+      :release_date => '30-May-2019'},
+    {:title => 'Nomadland', :rating => 'R',
+      :release_date => '19-Feb-2021'},
+    {:title => 'CODA', :rating => 'PG-13',
+      :release_date => '13-Aug-2021'}
+  ]
+  
+  more_movies.each do |movie|
+    Movie.create!(movie)
+  end
